+++ backend/src/main/java/com/vicenzo/customer/repository/CustomerDao.java
@@ -8,15 +8,15 @@ import java.util.Optional;
 public interface CustomerDao {
     List<Customer> selectAllCustomers();
 
-    Optional<Customer> selectCustomerById(Integer customerId);
+    Optional<Customer> selectCustomerById(Long customerId);
 
     void insertCustomer(Customer customer);
 
     boolean existsPersonWithEmail(String email);
 
-    boolean existsPersonWithId(Integer id);
+    boolean existsPersonWithId(Long id);
 
-    void deleteCustomerCustomerById(Integer id);
+    void deleteCustomerCustomerById(Long id);
 
     void updateCustomer(Customer customer);
 }
