+++ backend/src/main/java/com/vicenzo/customer/repository/CustomerRepository.java
@@ -3,8 +3,8 @@ package com.vicenzo.customer.repository;
 import com.vicenzo.customer.model.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
     boolean existsCustomerByEmail(String email);
 
-    boolean existsCustomerById(Integer id);
+    boolean existsCustomerById(Long id);
 }
